+++ Vali-Mediator/Core/Request/Unit.cs
@@ -9,8 +9,12 @@ public readonly struct Unit : IEquatable<Unit>
     /// <summary>The singleton value of <see cref="Unit"/>.</summary>
     public static readonly Unit Value = new();
 
+    /// <inheritdoc/>
     public bool Equals(Unit other) => true;
+    /// <inheritdoc/>
     public override bool Equals(object? obj) => obj is Unit;
+    /// <inheritdoc/>
     public override int GetHashCode() => 0;
+    /// <inheritdoc/>
     public override string ToString() => "()";
 }